Splitting tests into another package to avoid race condition
